@@ -2,7 +2,7 @@ name: media_kit
 description: A cross-platform video player & audio player for Flutter & Dart. Performant, stable, feature-proof & modular.
 homepage: https://github.com/media-kit/media-kit
 repository: https://github.com/media-kit/media-kit
-version: 1.1.10+1
+version: 1.1.11
 
 screenshots:
   - description: "package:media_kit running on Android (visible controls)"
@@ -29,24 +29,24 @@ topics:
   - cross-platform
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.1.0 <4.0.0"
 
 dependencies:
+  collection: ^1.17.0
+  http: ">=0.13.0 <2.0.0"
+  image: ^4.0.17
+  js: ^0.7.1
   meta: ^1.8.0
   path: ^1.8.0
-  image: ^4.0.17
-  uri_parser: ^2.0.2
-  collection: ^1.17.0
+  safe_local_storage: ^2.0.1
   synchronized: ^3.1.0
-  uuid: ">=2.0.0 <5.0.0"
-  http: ">=0.13.0 <2.0.0"
-  safe_local_storage: ^1.0.2
   universal_platform: ^1.0.0+1
-  js: ^0.6.7
+  uri_parser: ^3.0.0
+  uuid: ">=2.0.0 <5.0.0"
 
 dev_dependencies:
   test: ^1.24.1
-  lints: ^2.1.1
+  lints: ^5.0.0
 
 ffigen:
   name: MPV
