@@ -16,28 +16,25 @@ import 'package:safe_local_storage/safe_local_storage.dart';
 
 import 'package:media_kit/ffi/ffi.dart';
 
-import 'package:media_kit/src/player/platform_player.dart';
-
-import 'package:media_kit/src/player/native/core/initializer.dart';
-import 'package:media_kit/src/player/native/core/native_library.dart';
-import 'package:media_kit/src/player/native/core/fallback_bitrate_handler.dart';
-import 'package:media_kit/src/player/native/core/initializer_native_event_loop.dart';
-
-import 'package:media_kit/src/player/native/utils/isolates.dart';
-import 'package:media_kit/src/player/native/utils/temp_file.dart';
-import 'package:media_kit/src/player/native/utils/android_helper.dart';
-import 'package:media_kit/src/player/native/utils/android_asset_loader.dart';
-
-import 'package:media_kit/src/models/track.dart';
-import 'package:media_kit/src/models/playable.dart';
-import 'package:media_kit/src/models/playlist.dart';
-import 'package:media_kit/src/models/player_log.dart';
-import 'package:media_kit/src/models/media/media.dart';
 import 'package:media_kit/src/models/audio_device.dart';
 import 'package:media_kit/src/models/audio_params.dart';
-import 'package:media_kit/src/models/video_params.dart';
+import 'package:media_kit/src/models/media/media.dart';
+import 'package:media_kit/src/models/playable.dart';
+import 'package:media_kit/src/models/player_log.dart';
 import 'package:media_kit/src/models/player_state.dart';
 import 'package:media_kit/src/models/playlist_mode.dart';
+import 'package:media_kit/src/models/playlist.dart';
+import 'package:media_kit/src/models/track.dart';
+import 'package:media_kit/src/models/video_params.dart';
+import 'package:media_kit/src/player/native/core/fallback_bitrate_handler.dart';
+import 'package:media_kit/src/player/native/core/initializer.dart';
+import 'package:media_kit/src/player/native/core/native_library.dart';
+import 'package:media_kit/src/player/native/utils/android_asset_loader.dart';
+import 'package:media_kit/src/player/native/utils/android_helper.dart';
+import 'package:media_kit/src/player/native/utils/isolates.dart';
+import 'package:media_kit/src/player/native/utils/native_reference_holder.dart';
+import 'package:media_kit/src/player/native/utils/temp_file.dart';
+import 'package:media_kit/src/player/platform_player.dart';
 
 import 'package:media_kit/generated/libmpv/bindings.dart' as generated;
 
@@ -45,7 +42,25 @@ import 'package:media_kit/generated/libmpv/bindings.dart' as generated;
 void nativeEnsureInitialized({String? libmpv}) {
   AndroidHelper.ensureInitialized();
   NativeLibrary.ensureInitialized(libmpv: libmpv);
-  InitializerNativeEventLoop.ensureInitialized();
+  NativeReferenceHolder.ensureInitialized((references) async {
+    if (references.isEmpty) {
+      return;
+    }
+    const tag = NativeReferenceHolder.kTag;
+    print('$tag Found ${references.length} reference(s).');
+    print('$tag Disposing:\n${references.map((e) => e.address).join('\n')}');
+
+    // I can only get quit to work; [mpv_terminate_destroy] causes direct crash.
+    final mpv = generated.MPV(DynamicLibrary.open(NativeLibrary.path));
+    final cmd = 'quit'.toNativeUtf8();
+    try {
+      for (final reference in references) {
+        mpv.mpv_command_string(reference.cast(), cmd.cast());
+      }
+    } finally {
+      calloc.free(cmd);
+    }
+  });
 }
 
 /// {@template native_player}
@@ -78,17 +93,14 @@ class NativePlayer extends PlatformPlayer {
       await waitForPlayerInitialization;
       await waitForVideoControllerInitializationIfAttached;
 
-      await pause(synchronized: false);
-
-      await setVideoTrack(VideoTrack.no(), synchronized: false);
-      await setAudioTrack(AudioTrack.no(), synchronized: false);
-      await setSubtitleTrack(SubtitleTrack.no(), synchronized: false);
+      await NativeReferenceHolder.instance.remove(ctx);
+      await stop(notify: false, synchronized: false);
 
       disposed = true;
 
       await super.dispose();
 
-      Initializer.dispose(ctx);
+      Initializer(mpv).dispose(ctx);
 
       Future.delayed(const Duration(seconds: 5), () {
         mpv.mpv_terminate_destroy(ctx);
@@ -148,23 +160,6 @@ class NativePlayer extends PlatformPlayer {
       // Keep these [Media] objects in memory.
       current = playlist;
 
-      // NOTE: Handled as part of [stop] logic.
-      // final commands = [
-      //   // Clear existing playlist & change currently playing index to none.
-      //   // This causes playback to stop & player to enter the idle state.
-      //   'stop',
-      //   'playlist-clear',
-      //   'playlist-play-index none',
-      // ];
-      // for (final command in commands) {
-      //   final args = command.toNativeUtf8();
-      //   mpv.mpv_command_string(
-      //     ctx,
-      //     args.cast(),
-      //   );
-      //   calloc.free(args);
-      // }
-
       // Restore original state & reset public [PlayerState] & [PlayerStream] values e.g. width=null, height=null, subtitle=['', ''] etc.
       await stop(
         open: true,
@@ -172,68 +167,48 @@ class NativePlayer extends PlatformPlayer {
       );
 
       // Enter paused state.
-      {
-        final name = 'pause'.toNativeUtf8();
-        final value = calloc<Uint8>();
-        mpv.mpv_get_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_FLAG,
-          value.cast(),
-        );
-        if (value.value == 0) {
-          // We are using `cycle pause` because it waits & prevents the race condition.
-          final command = 'cycle pause'.toNativeUtf8();
-          mpv.mpv_command_string(
-            ctx,
-            command.cast(),
+      await _setPropertyFlag('pause', true);
+
+      if (playlist.any((media) => media.uri.startsWith('fd://'))) {
+        // `fd://` is a scheme used to reference `content` URIs on Android,
+        // but the `loadlist` command does not support that scheme by default,
+        // yielding "Refusing to load potentially unsafe URL from a playlist."
+        // so we fallback to loading each file individually.
+        for (int i = 0; i < playlist.length; i++) {
+          await _command(
+            [
+              'loadfile',
+              playlist[i].uri,
+              'append',
+            ],
           );
-          // NOTE: Handled as part of [stop] logic.
-          // state = state.copyWith(playing: false);
-          // if (!playingController.isClosed) {
-          //   playingController.add(false);
-          // }
+        }
+      } else {
+        final file = await TempFile.create();
+        String list = '';
+        for (final media in playlist) {
+          list += '${media.uri}\n';
         }
 
-        calloc.free(name);
-        calloc.free(value);
-      }
+        await file.write_(list);
 
-      // NOTE: Handled as part of [stop] logic.
-      // isShuffleEnabled = false;
-      // isPlayingStateChangeAllowed = false;
-
-      for (int i = 0; i < playlist.length; i++) {
         await _command(
           [
-            'loadfile',
-            playlist[i].uri,
+            'loadlist',
+            file.path,
             'append',
           ],
         );
+
+        Future.delayed(const Duration(seconds: 5), () {
+          file.delete_();
+        });
       }
 
       // If [play] is `true`, then exit paused state.
       if (play) {
         isPlayingStateChangeAllowed = true;
-        final name = 'pause'.toNativeUtf8();
-        final value = calloc<Uint8>();
-        mpv.mpv_get_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_FLAG,
-          value.cast(),
-        );
-        if (value.value == 1) {
-          // We are using `cycle pause` because it waits & prevents the race condition.
-          final command = 'cycle pause'.toNativeUtf8();
-          mpv.mpv_command_string(
-            ctx,
-            command.cast(),
-          );
-        }
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyFlag('pause', false);
         state = state.copyWith(playing: true);
         if (!playingController.isClosed) {
           playingController.add(true);
@@ -241,18 +216,7 @@ class NativePlayer extends PlatformPlayer {
       }
 
       // Jump to the specified [index] (in both cases either [play] is `true` or `false`).
-      {
-        final name = 'playlist-pos'.toNativeUtf8();
-        final value = calloc<Int64>()..value = index;
-        mpv.mpv_set_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_INT64,
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
-      }
+      await _setPropertyInt64('playlist-pos', index);
     }
 
     if (synchronized) {
@@ -267,6 +231,7 @@ class NativePlayer extends PlatformPlayer {
   @override
   Future<void> stop({
     bool open = false,
+    bool notify = true,
     bool synchronized = true,
   }) async {
     Future<void> function() async {
@@ -281,17 +246,12 @@ class NativePlayer extends PlatformPlayer {
       isBufferingStateChangeAllowed = false;
 
       final commands = [
-        'stop',
-        'playlist-clear',
-        'playlist-play-index none',
+        ['stop'],
+        ['playlist-clear'],
+        ['playlist-play-index', 'none'],
       ];
       for (final command in commands) {
-        final args = command.toNativeUtf8();
-        mpv.mpv_command_string(
-          ctx,
-          args.cast(),
-        );
-        calloc.free(args);
+        await _command(command);
       }
 
       // Reset the remaining attributes.
@@ -303,71 +263,73 @@ class NativePlayer extends PlatformPlayer {
         audioDevice: state.audioDevice,
         audioDevices: state.audioDevices,
       );
-      if (!open) {
-        // Do not emit PlayerStream.playlist if invoked from [open].
-        if (!playlistController.isClosed) {
-          playlistController.add(Playlist([]));
+      if (notify) {
+        if (!open) {
+          // Do not emit PlayerStream.playlist if invoked from [open].
+          if (!playlistController.isClosed) {
+            playlistController.add(Playlist([]));
+          }
         }
-      }
-      if (!playingController.isClosed) {
-        playingController.add(false);
-      }
-      if (!completedController.isClosed) {
-        completedController.add(false);
-      }
-      if (!positionController.isClosed) {
-        positionController.add(Duration.zero);
-      }
-      if (!durationController.isClosed) {
-        durationController.add(Duration.zero);
-      }
-      // if (!volumeController.isClosed) {
-      //   volumeController.add(0.0);
-      // }
-      // if (!rateController.isClosed) {
-      //   rateController.add(0.0);
-      // }
-      // if (!pitchController.isClosed) {
-      //   pitchController.add(0.0);
-      // }
-      if (!bufferingController.isClosed) {
-        bufferingController.add(false);
-      }
-      if (!bufferController.isClosed) {
-        bufferController.add(Duration.zero);
-      }
-      // if (!playlistModeController.isClosed) {
-      //   playlistModeController.add(PlaylistMode.none);
-      // }
-      if (!audioParamsController.isClosed) {
-        audioParamsController.add(const AudioParams());
-      }
-      if (!videoParamsController.isClosed) {
-        videoParamsController.add(const VideoParams());
-      }
-      if (!audioBitrateController.isClosed) {
-        audioBitrateController.add(null);
-      }
-      // if (!audioDeviceController.isClosed) {
-      //   audioDeviceController.add(AudioDevice.auto());
-      // }
-      // if (!audioDevicesController.isClosed) {
-      //   audioDevicesController.add([AudioDevice.auto()]);
-      // }
-      if (!trackController.isClosed) {
-        trackController.add(Track());
-      }
-      if (!tracksController.isClosed) {
-        tracksController.add(Tracks());
-      }
-      if (!widthController.isClosed) {
-        widthController.add(null);
-      }
-      if (!heightController.isClosed) {
-        heightController.add(null);
-      }
-      if (!subtitleController.isClosed) {
-        subtitleController.add(['', '']);
+        if (!playingController.isClosed) {
+          playingController.add(false);
+        }
+        if (!completedController.isClosed) {
+          completedController.add(false);
+        }
+        if (!positionController.isClosed) {
+          positionController.add(Duration.zero);
+        }
+        if (!durationController.isClosed) {
+          durationController.add(Duration.zero);
+        }
+        // if (!volumeController.isClosed) {
+        //   volumeController.add(0.0);
+        // }
+        // if (!rateController.isClosed) {
+        //   rateController.add(0.0);
+        // }
+        // if (!pitchController.isClosed) {
+        //   pitchController.add(0.0);
+        // }
+        if (!bufferingController.isClosed) {
+          bufferingController.add(false);
+        }
+        if (!bufferController.isClosed) {
+          bufferController.add(Duration.zero);
+        }
+        // if (!playlistModeController.isClosed) {
+        //   playlistModeController.add(PlaylistMode.none);
+        // }
+        // if (!audioParamsController.isClosed) {
+        //   audioParamsController.add(const AudioParams());
+        // }
+        if (!videoParamsController.isClosed) {
+          videoParamsController.add(const VideoParams());
+        }
+        // if (!audioBitrateController.isClosed) {
+        //   audioBitrateController.add(null);
+        // }
+        // if (!audioDeviceController.isClosed) {
+        //   audioDeviceController.add(AudioDevice.auto());
+        // }
+        // if (!audioDevicesController.isClosed) {
+        //   audioDevicesController.add([AudioDevice.auto()]);
+        // }
+        if (!trackController.isClosed) {
+          trackController.add(Track());
+        }
+        if (!tracksController.isClosed) {
+          tracksController.add(Tracks());
+        }
+        if (!widthController.isClosed) {
+          widthController.add(null);
+        }
+        if (!heightController.isClosed) {
+          heightController.add(null);
+        }
+        if (!subtitleController.isClosed) {
+          subtitleController.add(['', '']);
+        }
       }
     }
 
@@ -393,22 +355,14 @@ class NativePlayer extends PlatformPlayer {
         playingController.add(true);
       }
 
-      final name = 'pause'.toNativeUtf8();
-      final value = calloc<Uint8>();
-      mpv.mpv_get_property(
-        ctx,
-        name.cast(),
-        generated.mpv_format.MPV_FORMAT_FLAG,
-        value.cast(),
-      );
-      if (value.value == 1) {
-        await playOrPause(
-          notify: false,
-          synchronized: false,
-        );
+      isPlayingStateChangeAllowed = true;
+
+      // This condition is specifically for the case when the internal playlist is ended (with [PlaylistLoopMode.none]), and we want to play the playlist again if play/pause is pressed.
+      if (state.completed) {
+        await seek(Duration.zero, synchronized: false);
+        await _setPropertyInt64('playlist-pos', 0);
       }
-      calloc.free(name);
-      calloc.free(value);
+      await _setPropertyFlag('pause', false);
     }
 
     if (synchronized) {
@@ -433,22 +387,9 @@ class NativePlayer extends PlatformPlayer {
         playingController.add(false);
       }
 
-      final name = 'pause'.toNativeUtf8();
-      final value = calloc<Uint8>();
-      mpv.mpv_get_property(
-        ctx,
-        name.cast(),
-        generated.mpv_format.MPV_FORMAT_FLAG,
-        value.cast(),
-      );
-      if (value.value == 0) {
-        await playOrPause(
-          notify: false,
-          synchronized: false,
-        );
-      }
-      calloc.free(name);
-      calloc.free(value);
+      isPlayingStateChangeAllowed = true;
+      isBufferingStateChangeAllowed = false;
+      await _setPropertyFlag('pause', true);
     }
 
     if (synchronized) {
@@ -487,23 +428,9 @@ class NativePlayer extends PlatformPlayer {
       // This condition is specifically for the case when the internal playlist is ended (with [PlaylistLoopMode.none]), and we want to play the playlist again if play/pause is pressed.
       if (state.completed) {
         await seek(Duration.zero, synchronized: false);
-        final name = 'playlist-pos'.toNativeUtf8();
-        final value = calloc<Int64>()..value = 0;
-        mpv.mpv_set_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_INT64,
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyInt64('playlist-pos', 0);
       }
-      final command = 'cycle pause'.toNativeUtf8();
-      mpv.mpv_command_string(
-        ctx,
-        command.cast(),
-      );
-      calloc.free(command);
+      await _command(['cycle', 'pause']);
     }
 
     if (synchronized) {
@@ -528,12 +455,7 @@ class NativePlayer extends PlatformPlayer {
       current.add(media);
       // ---------------------------------------------
 
-      final command = 'loadfile ${media.uri} append'.toNativeUtf8();
-      mpv.mpv_command_string(
-        ctx,
-        command.cast(),
-      );
-      calloc.free(command.cast());
+      await _command(['loadfile', media.uri, 'append']);
     }
 
     if (synchronized) {
@@ -587,12 +509,7 @@ class NativePlayer extends PlatformPlayer {
         }
       }
 
-      final command = 'playlist-remove $index'.toNativeUtf8();
-      mpv.mpv_command_string(
-        ctx,
-        command.cast(),
-      );
-      calloc.free(command.cast());
+      await _command(['playlist-remove', index.toString()]);
     }
 
     if (synchronized) {
@@ -622,12 +539,7 @@ class NativePlayer extends PlatformPlayer {
       }
 
       await play(synchronized: false);
-      final command = 'playlist-next'.toNativeUtf8();
-      mpv.mpv_command_string(
-        ctx,
-        command.cast(),
-      );
-      calloc.free(command);
+      await _command(['playlist-next']);
     }
 
     if (synchronized) {
@@ -657,12 +569,7 @@ class NativePlayer extends PlatformPlayer {
       }
 
       await play(synchronized: false);
-      final command = 'playlist-prev'.toNativeUtf8();
-      mpv.mpv_command_string(
-        ctx,
-        command.cast(),
-      );
-      calloc.free(command);
+      await _command(['playlist-prev']);
     }
 
     if (synchronized) {
@@ -683,16 +590,7 @@ class NativePlayer extends PlatformPlayer {
       await waitForVideoControllerInitializationIfAttached;
 
       await play(synchronized: false);
-      final name = 'playlist-pos'.toNativeUtf8();
-      final value = calloc<Int64>()..value = index;
-      mpv.mpv_set_property(
-        ctx,
-        name.cast(),
-        generated.mpv_format.MPV_FORMAT_INT64,
-        value.cast(),
-      );
-      calloc.free(name);
-      calloc.free(value);
+      await _setPropertyInt64('playlist-pos', index);
     }
 
     if (synchronized) {
@@ -725,12 +623,7 @@ class NativePlayer extends PlatformPlayer {
       current = values;
       // ---------------------------------------------
 
-      final command = 'playlist-move $from $to'.toNativeUtf8();
-      mpv.mpv_command_string(
-        ctx,
-        command.cast(),
-      );
-      calloc.free(command.cast());
+      await _command(['playlist-move', from.toString(), to.toString()]);
     }
 
     if (synchronized) {
@@ -751,14 +644,11 @@ class NativePlayer extends PlatformPlayer {
       await waitForPlayerInitialization;
       await waitForVideoControllerInitializationIfAttached;
 
-      await compute(
-        _seek,
-        _SeekData(
-          ctx.address,
-          NativeLibrary.path,
-          duration,
-        ),
-      );
+      await _command([
+        'seek',
+        (duration.inMilliseconds / 1000).toStringAsFixed(4),
+        'absolute'
+      ]);
 
       // It is self explanatory that PlayerState.completed & PlayerStream.completed must enter the false state if seek is called. Typically after EOF.
       // https://github.com/media-kit/media-kit/issues/221
@@ -786,60 +676,28 @@ class NativePlayer extends PlatformPlayer {
       await waitForPlayerInitialization;
       await waitForVideoControllerInitializationIfAttached;
 
-      final file = 'loop-file'.toNativeUtf8();
-      final playlist = 'loop-playlist'.toNativeUtf8();
-      final yes = 'yes'.toNativeUtf8();
-      final no = 'no'.toNativeUtf8();
       switch (playlistMode) {
         case PlaylistMode.none:
           {
-            mpv.mpv_set_property_string(
-              ctx,
-              file.cast(),
-              no.cast(),
-            );
-            mpv.mpv_set_property_string(
-              ctx,
-              playlist.cast(),
-              no.cast(),
-            );
+            await _setPropertyString('loop-file', 'no');
+            await _setPropertyString('loop-playlist', 'no');
             break;
           }
         case PlaylistMode.single:
           {
-            mpv.mpv_set_property_string(
-              ctx,
-              file.cast(),
-              yes.cast(),
-            );
-            mpv.mpv_set_property_string(
-              ctx,
-              playlist.cast(),
-              no.cast(),
-            );
+            await _setPropertyString('loop-file', 'yes');
+            await _setPropertyString('loop-playlist', 'no');
             break;
           }
         case PlaylistMode.loop:
           {
-            mpv.mpv_set_property_string(
-              ctx,
-              file.cast(),
-              no.cast(),
-            );
-            mpv.mpv_set_property_string(
-              ctx,
-              playlist.cast(),
-              yes.cast(),
-            );
+            await _setPropertyString('loop-file', 'no');
+            await _setPropertyString('loop-playlist', 'yes');
             break;
           }
         default:
           break;
       }
-      calloc.free(file);
-      calloc.free(playlist);
-      calloc.free(yes);
-      calloc.free(no);
 
       state = state.copyWith(playlistMode: playlistMode);
       if (!playlistModeController.isClosed) {
@@ -864,40 +722,7 @@ class NativePlayer extends PlatformPlayer {
       await waitForPlayerInitialization;
       await waitForVideoControllerInitializationIfAttached;
 
-      {
-        final name = 'mute'.toNativeUtf8();
-        final value = calloc<Bool>();
-        mpv.mpv_get_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_FLAG,
-          value.cast(),
-        );
-        if (value.value) {
-          // Unmute the player before setting the volume.
-          final command = 'cycle mute'.toNativeUtf8();
-          mpv.mpv_command_string(
-            ctx,
-            command.cast(),
-          );
-          calloc.free(command);
-        }
-        calloc.free(name);
-        calloc.free(value);
-      }
-      {
-        final name = 'volume'.toNativeUtf8();
-        final value = calloc<Double>();
-        value.value = volume;
-        mpv.mpv_set_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_DOUBLE,
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
-      }
+      await _setPropertyDouble('volume', volume);
     }
 
     if (synchronized) {
@@ -937,27 +762,10 @@ class NativePlayer extends PlatformPlayer {
         // Apparently, using scaletempo:scale actually controls the playback rate as intended after setting audio-pitch-correction as FALSE.
         // speed on the other hand, changes the pitch when audio-pitch-correction is set to FALSE.
         // Since, it also alters the actual [speed], the scaletempo:scale is divided by the same value of [pitch] to compensate the speed change.
-        var name = 'audio-pitch-correction'.toNativeUtf8();
-        final no = 'no'.toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          no.cast(),
-        );
-        calloc.free(name);
-        calloc.free(no);
-        name = 'af'.toNativeUtf8();
+        await _setPropertyFlag('audio-pitch-correction', false);
         // Divide by [state.pitch] to compensate the speed change caused by pitch shift.
-        final value =
-            'scaletempo:scale=${(state.rate / state.pitch).toStringAsFixed(8)}'
-                .toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyString('af',
+            'scaletempo:scale=${(state.rate / state.pitch).toStringAsFixed(8)}');
       } else {
         // Pitch shift control is disabled.
 
@@ -967,17 +775,7 @@ class NativePlayer extends PlatformPlayer {
         if (!rateController.isClosed) {
           rateController.add(state.rate);
         }
-        final name = 'speed'.toNativeUtf8();
-        final value = calloc<Double>();
-        value.value = rate;
-        mpv.mpv_set_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_DOUBLE,
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyDouble('speed', rate);
       }
     }
 
@@ -1018,37 +816,11 @@ class NativePlayer extends PlatformPlayer {
         // Apparently, using scaletempo:scale actually controls the playback rate as intended after setting audio-pitch-correction as FALSE.
         // speed on the other hand, changes the pitch when audio-pitch-correction is set to FALSE.
         // Since, it also alters the actual [speed], the scaletempo:scale is divided by the same value of [pitch] to compensate the speed change.
-        var name = 'audio-pitch-correction'.toNativeUtf8();
-        final no = 'no'.toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          no.cast(),
-        );
-        calloc.free(name);
-        calloc.free(no);
-        name = 'speed'.toNativeUtf8();
-        final speed = calloc<Double>()..value = pitch;
-        mpv.mpv_set_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_DOUBLE,
-          speed.cast(),
-        );
-        calloc.free(name);
-        calloc.free(speed);
-        name = 'af'.toNativeUtf8();
+        await _setPropertyFlag('audio-pitch-correction', false);
+        await _setPropertyDouble('speed', pitch);
         // Divide by [state.pitch] to compensate the speed change caused by pitch shift.
-        final value =
-            'scaletempo:scale=${(state.rate / state.pitch).toStringAsFixed(8)}'
-                .toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyString('af',
+            'scaletempo:scale=${(state.rate / state.pitch).toStringAsFixed(8)}');
       } else {
         // Pitch shift control is disabled.
         throw ArgumentError('[PlayerConfiguration.pitch] is false');
@@ -1105,15 +877,7 @@ class NativePlayer extends PlatformPlayer {
       await waitForPlayerInitialization;
       await waitForVideoControllerInitializationIfAttached;
 
-      final name = 'audio-device'.toNativeUtf8();
-      final value = audioDevice.name.toNativeUtf8();
-      mpv.mpv_set_property_string(
-        ctx,
-        name.cast(),
-        value.cast(),
-      );
-      calloc.free(name);
-      calloc.free(value);
+      await _setPropertyString('audio-device', audioDevice.name);
     }
 
     if (synchronized) {
@@ -1136,15 +900,7 @@ class NativePlayer extends PlatformPlayer {
       await waitForPlayerInitialization;
       await waitForVideoControllerInitializationIfAttached;
 
-      final name = 'vid'.toNativeUtf8();
-      final value = track.id.toNativeUtf8();
-      mpv.mpv_set_property_string(
-        ctx,
-        name.cast(),
-        value.cast(),
-      );
-      calloc.free(name);
-      calloc.free(value);
+      await _setPropertyString('vid', track.id);
       state = state.copyWith(
         track: state.track.copyWith(
           video: track,
@@ -1206,15 +962,7 @@ class NativePlayer extends PlatformPlayer {
           trackController.add(state.track);
         }
       } else {
-        final name = 'aid'.toNativeUtf8();
-        final value = track.id.toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyString('aid', track.id);
         state = state.copyWith(
           track: state.track.copyWith(
             audio: track,
@@ -1300,15 +1048,7 @@ class NativePlayer extends PlatformPlayer {
           trackController.add(state.track);
         }
       } else {
-        final name = 'sid'.toNativeUtf8();
-        final value = track.id.toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyString('sid', track.id);
         state = state.copyWith(
           track: state.track.copyWith(
             subtitle: track,
@@ -1333,9 +1073,14 @@ class NativePlayer extends PlatformPlayer {
   /// * `image/jpeg`: Returns a JPEG encoded image.
   /// * `image/png`: Returns a PNG encoded image.
   /// * `null`: Returns BGRA pixel buffer.
+  ///
+  /// If [includeLibassSubtitles] is `true` *and* [PlayerConfiguration.libass] is `true`, then the
+  /// screenshot will include the on-screen subtitles.
   @override
   Future<Uint8List?> screenshot(
-      {String? format = 'image/jpeg', bool synchronized = true}) async {
+      {String? format = 'image/jpeg',
+      bool synchronized = true,
+      bool includeLibassSubtitles = false}) async {
     Future<Uint8List?> function() async {
       if (![
         'image/jpeg',
@@ -1361,6 +1106,7 @@ class NativePlayer extends PlatformPlayer {
           ctx.address,
           NativeLibrary.path,
           format,
+          includeLibassSubtitles,
         ),
       );
     }
@@ -1389,12 +1135,19 @@ class NativePlayer extends PlatformPlayer {
   /// * https://mpv.io/manual/master/#options
   /// * https://mpv.io/manual/master/#properties
   ///
-  Future<void> setProperty(String property, String value) async {
+  Future<void> setProperty(
+    String property,
+    String value, {
+    bool waitForInitialization = true,
+  }) async {
     if (disposed) {
       throw AssertionError('[Player] has been disposed');
     }
-    await waitForPlayerInitialization;
-    await waitForVideoControllerInitializationIfAttached;
+
+    if (waitForInitialization) {
+      await waitForPlayerInitialization;
+      await waitForVideoControllerInitializationIfAttached;
+    }
 
     final name = property.toNativeUtf8();
     final data = value.toNativeUtf8();
@@ -1414,12 +1167,18 @@ class NativePlayer extends PlatformPlayer {
   /// * https://mpv.io/manual/master/#options
   /// * https://mpv.io/manual/master/#properties
   ///
-  Future<String> getProperty(String property) async {
+  Future<String> getProperty(
+    String property, {
+    bool waitForInitialization = true,
+  }) async {
     if (disposed) {
       throw AssertionError('[Player] has been disposed');
     }
-    await waitForPlayerInitialization;
-    await waitForVideoControllerInitializationIfAttached;
+
+    if (waitForInitialization) {
+      await waitForPlayerInitialization;
+      await waitForVideoControllerInitializationIfAttached;
+    }
 
     final name = property.toNativeUtf8();
     final value = mpv.mpv_get_property_string(ctx, name.cast());
@@ -1443,13 +1202,17 @@ class NativePlayer extends PlatformPlayer {
   ///
   Future<void> observeProperty(
     String property,
-    Future<void> Function(String) listener,
-  ) async {
+    Future<void> Function(String) listener, {
+    bool waitForInitialization = true,
+  }) async {
     if (disposed) {
       throw AssertionError('[Player] has been disposed');
     }
-    await waitForPlayerInitialization;
-    await waitForVideoControllerInitializationIfAttached;
+
+    if (waitForInitialization) {
+      await waitForPlayerInitialization;
+      await waitForVideoControllerInitializationIfAttached;
+    }
 
     if (observed.containsKey(property)) {
       throw ArgumentError.value(
@@ -1477,12 +1240,18 @@ class NativePlayer extends PlatformPlayer {
   /// * https://mpv.io/manual/master/#options
   /// * https://mpv.io/manual/master/#properties
   ///
-  Future<void> unobserveProperty(String property) async {
+  Future<void> unobserveProperty(
+    String property, {
+    bool waitForInitialization = true,
+  }) async {
     if (disposed) {
       throw AssertionError('[Player] has been disposed');
     }
-    await waitForPlayerInitialization;
-    await waitForVideoControllerInitializationIfAttached;
+
+    if (waitForInitialization) {
+      await waitForPlayerInitialization;
+      await waitForVideoControllerInitializationIfAttached;
+    }
 
     if (!observed.containsKey(property)) {
       throw ArgumentError.value(
@@ -1502,12 +1271,18 @@ class NativePlayer extends PlatformPlayer {
   /// See:
   /// * https://mpv.io/manual/master/#list-of-input-commands
   ///
-  Future<void> command(List<String> command) async {
+  Future<void> command(
+    List<String> command, {
+    bool waitForInitialization = true,
+  }) async {
     if (disposed) {
       throw AssertionError('[Player] has been disposed');
     }
-    await waitForPlayerInitialization;
-    await waitForVideoControllerInitializationIfAttached;
+
+    if (waitForInitialization) {
+      await waitForPlayerInitialization;
+      await waitForVideoControllerInitializationIfAttached;
+    }
 
     await _command(command);
   }
@@ -1575,13 +1350,35 @@ class NativePlayer extends PlatformPlayer {
         }
       }
     }
+    if (event.ref.event_id ==
+        generated.mpv_event_id.MPV_EVENT_SET_PROPERTY_REPLY) {
+      final completer = _setPropertyRequests.remove(event.ref.reply_userdata);
+      if (completer == null) {
+        print(
+            'Warning: Received MPV_EVENT_SET_PROPERTY_REPLY with unregistered ID ${event.ref.reply_userdata}');
+      } else {
+        completer.complete(event.ref.error);
+      }
+    }
+    if (event.ref.event_id == generated.mpv_event_id.MPV_EVENT_COMMAND_REPLY) {
+      final completer = _commandRequests.remove(event.ref.reply_userdata);
+      if (completer == null) {
+        print(
+            'Warning: Received MPV_EVENT_COMMAND_REPLY with unregistered ID ${event.ref.reply_userdata}');
+      } else {
+        completer.complete(event.ref.error);
+      }
+    }
 
     if (!completer.isCompleted) {
       // Ignore the events which are fired before the initialization.
       return;
     }
 
-    _error(event.ref.error);
+    _logError(
+      event.ref.error,
+      'event:${event.ref.event_id} ${event.ref.data.cast<Uint8>()}',
+    );
 
     if (event.ref.event_id == generated.mpv_event_id.MPV_EVENT_START_FILE) {
       if (isPlayingStateChangeAllowed) {
@@ -1667,6 +1464,16 @@ class NativePlayer extends PlatformPlayer {
         state = state.copyWith(buffer: buffer);
         if (!bufferController.isClosed) {
           bufferController.add(buffer);
+        }
+      }
+      if (prop.ref.name.cast<Utf8>().toDartString() ==
+              'cache-buffering-state' &&
+          prop.ref.format == generated.mpv_format.MPV_FORMAT_DOUBLE) {
+        final bufferingPercentage = prop.ref.data.cast<Double>().value;
+
+        state = state.copyWith(bufferingPercentage: bufferingPercentage);
+        if (!bufferingPercentageController.isClosed) {
+          bufferingPercentageController.add(bufferingPercentage);
         }
       }
       if (prop.ref.name.cast<Utf8>().toDartString() == 'time-pos' &&
@@ -2204,7 +2011,12 @@ class NativePlayer extends PlatformPlayer {
           if (fn != null) {
             final data = mpv.mpv_get_property_string(ctx, prop.ref.name);
             if (data != nullptr) {
-              await fn.call(data.cast<Utf8>().toDartString());
+              try {
+                await fn.call(data.cast<Utf8>().toDartString());
+              } catch (exception, stacktrace) {
+                print(exception);
+                print(stacktrace);
+              }
               mpv.mpv_free(data.cast());
             }
           }
@@ -2353,7 +2165,9 @@ class NativePlayer extends PlatformPlayer {
             if (start != null) {
               try {
                 final property = 'start'.toNativeUtf8();
-                final value = (start.inMilliseconds / 1000).toStringAsFixed(3).toNativeUtf8();
+                final value = (start.inMilliseconds / 1000)
+                    .toStringAsFixed(3)
+                    .toNativeUtf8();
                 mpv.mpv_set_property_string(
                   ctx,
                   property.cast(),
@@ -2370,7 +2184,9 @@ class NativePlayer extends PlatformPlayer {
             if (end != null) {
               try {
                 final property = 'end'.toNativeUtf8();
-                final value = (end.inMilliseconds / 1000).toStringAsFixed(3).toNativeUtf8();
+                final value = (end.inMilliseconds / 1000)
+                    .toStringAsFixed(3)
+                    .toNativeUtf8();
                 mpv.mpv_set_property_string(
                   ctx,
                   property.cast(),
@@ -2471,7 +2287,8 @@ class NativePlayer extends PlatformPlayer {
 
       if (Platform.isAndroid &&
           configuration.libass &&
-          configuration.libassAndroidFont != null) {
+          configuration.libassAndroidFont != null &&
+          configuration.libassAndroidFontName != null) {
         try {
           // On Android, the system fonts cannot be picked up by libass/fontconfig. This makes libass subtitle rendering fail.
           // We save the subtitle font to the application's cache directory and set `config` & `config-dir` to use it.
@@ -2488,7 +2305,8 @@ class NativePlayer extends PlatformPlayer {
             options.addAll(
               {
                 'config': 'yes',
-                'config-dir': directory,
+                'sub-fonts-dir': directory,
+                'sub-font': configuration.libassAndroidFontName ?? '',
               },
             );
             print(subfont);
@@ -2500,8 +2318,7 @@ class NativePlayer extends PlatformPlayer {
         }
       }
 
-      ctx = await Initializer.create(
-        NativeLibrary.path,
+      ctx = await Initializer(mpv).create(
         _handler,
         options: options,
       );
@@ -2541,6 +2358,9 @@ class NativePlayer extends PlatformPlayer {
         'dscale': 'bilinear',
         'dither': 'no',
         'cache': 'yes',
+        'cache-on-disk': 'yes',
+        'hr-seek': 'yes',
+        'hr-seek-framedrop': 'no',
         'correct-downscaling': 'no',
         'linear-downscaling': 'no',
         'sigmoid-upscaling': 'no',
@@ -2582,29 +2402,11 @@ class NativePlayer extends PlatformPlayer {
         properties['ao'] = 'null';
       }
 
-      for (final property in properties.entries) {
-        final name = property.key.toNativeUtf8();
-        final value = property.value.toNativeUtf8();
-        mpv.mpv_set_property_string(
-          ctx,
-          name.cast(),
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
-      }
+      await Future.wait(properties.entries
+          .map((entry) => _setPropertyString(entry.key, entry.value)));
 
       if (configuration.muted) {
-        final name = 'mute'.toNativeUtf8();
-        final value = calloc<Bool>()..value = true;
-        mpv.mpv_set_property(
-          ctx,
-          name.cast(),
-          generated.mpv_format.MPV_FORMAT_FLAG,
-          value.cast(),
-        );
-        calloc.free(name);
-        calloc.free(value);
+        await _setPropertyDouble('volume', 0);
 
         state = state.copyWith(volume: 0.0);
         if (!volumeController.isClosed) {
@@ -2623,6 +2425,7 @@ class NativePlayer extends PlatformPlayer {
         'core-idle': generated.mpv_format.MPV_FORMAT_FLAG,
         'paused-for-cache': generated.mpv_format.MPV_FORMAT_FLAG,
         'demuxer-cache-time': generated.mpv_format.MPV_FORMAT_DOUBLE,
+        'cache-buffering-state': generated.mpv_format.MPV_FORMAT_DOUBLE,
         'audio-params': generated.mpv_format.MPV_FORMAT_NODE,
         'audio-bitrate': generated.mpv_format.MPV_FORMAT_DOUBLE,
         'audio-device': generated.mpv_format.MPV_FORMAT_NODE,
@@ -2669,15 +2472,95 @@ class NativePlayer extends PlatformPlayer {
       mpv.mpv_hook_add(ctx, 0, unload.cast(), 0);
       calloc.free(load);
       calloc.free(unload);
+
+      await NativeReferenceHolder.instance.add(ctx);
     });
   }
 
   /// Adds an error to the [Player.stream.error].
-  void _error(int code) {
-    if (code < 0 && !errorController.isClosed) {
+  void _logError(int code, String? text) {
+    if (code < 0 && !logController.isClosed) {
       final message = mpv.mpv_error_string(code).cast<Utf8>().toDartString();
-      errorController.add(message);
+      logController.add(PlayerLog(
+        prefix: 'media_kit',
+        level: 'error',
+        text: 'error: $message $text',
+      ));
     }
+  }
+
+  int _asyncRequestNumber = 0;
+  final Map<int, Completer<int>> _setPropertyRequests = {};
+  final Map<int, Completer<int>> _commandRequests = {};
+
+  Future<void> _setProperty(
+    String name,
+    int format,
+    Pointer<Void> data,
+  ) async {
+    final requestNumber = _asyncRequestNumber++;
+    final completer = _setPropertyRequests[requestNumber] = Completer<int>();
+    final namePtr = name.toNativeUtf8();
+    final immediate = mpv.mpv_set_property_async(
+      ctx,
+      requestNumber,
+      namePtr.cast(),
+      format,
+      data,
+    );
+    calloc.free(namePtr);
+    String text = '_setProperty($name, $format)';
+    if (immediate < 0) {
+      // Sending failed.
+      _logError(immediate, text);
+      return;
+    }
+
+    _logError(await completer.future, text);
+  }
+
+  Future<void> _setPropertyFlag(String name, bool value) async {
+    final ptr = calloc<Bool>(1)..value = value;
+    await _setProperty(
+      name,
+      generated.mpv_format.MPV_FORMAT_FLAG,
+      ptr.cast(),
+    );
+    calloc.free(ptr);
+  }
+
+  Future<void> _setPropertyDouble(String name, double value) async {
+    final ptr = calloc<Double>(1)..value = value;
+    await _setProperty(
+      name,
+      generated.mpv_format.MPV_FORMAT_DOUBLE,
+      ptr.cast(),
+    );
+    calloc.free(ptr);
+  }
+
+  Future<void> _setPropertyInt64(String name, int value) async {
+    final ptr = calloc<Int64>(1)..value = value;
+    await _setProperty(
+      name,
+      generated.mpv_format.MPV_FORMAT_INT64,
+      ptr.cast(),
+    );
+    calloc.free(ptr);
+  }
+
+  Future<void> _setPropertyString(String name, String value) async {
+    final string = value.toNativeUtf8();
+    // API requires char**.
+    final ptr = calloc<Pointer<Void>>(1);
+    ptr.value = Pointer.fromAddress(string.address);
+    await _setProperty(
+      name,
+      generated.mpv_format.MPV_FORMAT_STRING,
+      ptr.cast(),
+    );
+    calloc.free(ptr);
+    calloc.free(string);
   }
 
   /// Calls mpv command passed as [args].
@@ -2686,17 +2569,24 @@ class NativePlayer extends PlatformPlayer {
     final pointers = args.map<Pointer<Utf8>>((e) => e.toNativeUtf8()).toList();
     final arr = calloc<Pointer<Utf8>>(128);
     for (int i = 0; i < args.length; i++) {
-      arr.elementAt(i).value = pointers[i];
+      (arr + i).value = pointers[i];
     }
-    mpv.mpv_command(
-      ctx,
-      arr.cast(),
-    );
+    final requestNumber = _asyncRequestNumber++;
+    final completer = _commandRequests[requestNumber] = Completer<int>();
+    final immediate = mpv.mpv_command_async(ctx, requestNumber, arr.cast());
     calloc.free(arr);
     pointers.forEach(calloc.free);
+    String text = '_command(${args.join(', ')})';
+    if (immediate < 0) {
+      // Sending failed.
+      _logError(immediate, text);
+      return;
+    }
+
+    _logError(await completer.future, text);
   }
 
-  /// Internal generated libmpv C API bindings.
+  /// Generated libmpv C API bindings.
   final generated.MPV mpv;
 
   /// [Pointer] to [generated.mpv_handle] of this instance.
@@ -2761,44 +2651,17 @@ class NativePlayer extends PlatformPlayer {
 // TODO: Maybe eventually move all methods to [Isolate]?
 // --------------------------------------------------
 
-class _SeekData {
-  final int ctx;
-  final String lib;
-  final Duration duration;
-
-  _SeekData(
-    this.ctx,
-    this.lib,
-    this.duration,
-  );
-}
-
-/// [NativePlayer.seek]
-void _seek(_SeekData data) {
-  // ---------
-  final mpv = generated.MPV(DynamicLibrary.open(data.lib));
-  final ctx = Pointer<generated.mpv_handle>.fromAddress(data.ctx);
-  // ---------
-  final duration = data.duration;
-  // ---------
-  final value = duration.inMilliseconds / 1000;
-  final command = 'seek ${value.toStringAsFixed(4)} absolute'.toNativeUtf8();
-  mpv.mpv_command_string(
-    ctx,
-    command.cast(),
-  );
-  calloc.free(command);
-}
-
 class _ScreenshotData {
   final int ctx;
   final String lib;
   final String? format;
+  final bool includeLibassSubtitles;
 
   const _ScreenshotData(
     this.ctx,
     this.lib,
     this.format,
+    this.includeLibassSubtitles,
   );
 }
 
@@ -2814,7 +2677,7 @@ Uint8List? _screenshot(_ScreenshotData data) {
   // https://mpv.io/manual/stable/#command-interface-screenshot-raw
   final args = [
     'screenshot-raw',
-    'video',
+    data.includeLibassSubtitles ? 'subtitles' : 'video',
   ];
 
   final result = calloc<generated.mpv_node>();
