@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
-
 import 'package:media_kit/media_kit.dart';
 import 'package:media_kit_video/media_kit_video.dart';
 
 import '../common/globals.dart';
-import '../common/widgets.dart';
 import '../common/sources/sources.dart';
+import '../common/widgets.dart';
 
 class Screenshot extends StatefulWidget {
   const Screenshot({Key? key}) : super(key: key);
@@ -101,7 +100,6 @@ class _ScreenshotState extends State<Screenshot> {
                         children: [
                           Expanded(
                             child: Card(
-                              elevation: 8.0,
                               clipBehavior: Clip.antiAlias,
                               margin: const EdgeInsets.all(32.0),
                               child: Video(
