@@ -1,13 +1,12 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-
 import 'package:media_kit/media_kit.dart';
 import 'package:media_kit_video/media_kit_video.dart';
 
 import '../common/globals.dart';
-import '../common/widgets.dart';
 import '../common/sources/sources.dart';
+import '../common/widgets.dart';
 
 class ProgrammaticFullscreen extends StatefulWidget {
   const ProgrammaticFullscreen({Key? key}) : super(key: key);
@@ -132,7 +131,6 @@ class _ProgrammaticFullscreenState extends State<ProgrammaticFullscreen> {
                         children: [
                           Expanded(
                             child: Card(
-                              elevation: 8.0,
                               clipBehavior: Clip.antiAlias,
                               margin: const EdgeInsets.all(32.0),
                               child: Video(
