@@ -7,56 +7,26 @@
  */
 package com.alexmercerind.media_kit_video;
 
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.PorterDuff;
-import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
-import android.view.Surface;
-import android.view.View;
-import android.widget.FrameLayout;
 
-import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Locale;
+import java.util.Objects;
 
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.android.FlutterFragmentActivity;
-import io.flutter.embedding.android.FlutterView;
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.embedding.engine.FlutterJNI;
-import io.flutter.plugin.common.MethodChannel;
 import io.flutter.view.TextureRegistry;
 
+public class VideoOutput implements TextureRegistry.SurfaceProducer.Callback {
+    private static final String TAG = "VideoOutput";
+    private static final Method newGlobalObjectRef;
+    private static final Method deleteGlobalObjectRef;
+    private static final HashSet<Long> deletedGlobalObjectRefs = new HashSet<>();
+    private static final Handler handler = new Handler(Looper.getMainLooper());
 
-public class VideoOutput {
-    public long id = 0;
-    public long wid = 0;
-
-    private Surface surface;
-    private final TextureRegistry.SurfaceTextureEntry surfaceTextureEntry;
-
-    private boolean flutterJNIAPIAvailable;
-    private final Method newGlobalObjectRef;
-    private final Method deleteGlobalObjectRef;
-    private boolean waitUntilFirstFrameRenderedNotify;
-
-    private long handle;
-    private MethodChannel channelReference;
-    private TextureRegistry textureRegistryReference;
-
-    private final Object lock = new Object();
-
-    VideoOutput(long handle, MethodChannel channelReference, TextureRegistry textureRegistryReference) {
-        this.handle = handle;
-        this.channelReference = channelReference;
-        this.textureRegistryReference = textureRegistryReference;
+    static {
         try {
-            flutterJNIAPIAvailable = false;
-            waitUntilFirstFrameRenderedNotify = false;
             // com.alexmercerind.mediakitandroidhelper.MediaKitAndroidHelper is part of package:media_kit_libs_android_video & package:media_kit_libs_android_audio packages.
             // Use reflection to invoke methods of com.alexmercerind.mediakitandroidhelper.MediaKitAndroidHelper.
             Class<?> mediaKitAndroidHelperClass = Class.forName("com.alexmercerind.mediakitandroidhelper.MediaKitAndroidHelper");
@@ -68,181 +38,107 @@ public class VideoOutput {
             Log.i("media_kit", "package:media_kit_libs_android_video missing. Make sure you have added it to pubspec.yaml.");
             throw new RuntimeException("Failed to initialize com.alexmercerind.media_kit_video.VideoOutput.");
         }
+    }
 
-        surfaceTextureEntry = textureRegistryReference.createSurfaceTexture();
+    private long id = 0;
+    private long wid = 0;
 
-        // If we call setOnFrameAvailableListener after creating SurfaceTextureEntry, the texture won't be displayed inside Flutter UI, because callback set by us will override the Flutter engine's own registered callback:
-        // https://github.com/flutter/engine/blob/f47e864f2dcb9c299a3a3ed22300a1dcacbdf1fe/shell/platform/android/io/flutter/view/FlutterView.java#L942-L958
-        try {
-            if (!flutterJNIAPIAvailable) {
-                flutterJNIAPIAvailable = getFlutterJNIReference() != null;
-            }
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
-        Log.i("media_kit", String.format(Locale.ENGLISH, "flutterJNIAPIAvailable = %b", flutterJNIAPIAvailable));
-        if (flutterJNIAPIAvailable) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                surfaceTextureEntry.surfaceTexture().setOnFrameAvailableListener((texture) -> {
-                    synchronized (lock) {
-                        try {
-                            if (!waitUntilFirstFrameRenderedNotify) {
-                                waitUntilFirstFrameRenderedNotify = true;
-                                final HashMap<String, Object> data = new HashMap<>();
-                                data.put("handle", handle);
-                                channelReference.invokeMethod("VideoOutput.WaitUntilFirstFrameRenderedNotify", data);
-                                Log.i("media_kit", String.format(Locale.ENGLISH, "VideoOutput.WaitUntilFirstFrameRenderedNotify = %d", handle));
-                            }
+    private final TextureUpdateCallback textureUpdateCallback;
 
-                            FlutterJNI flutterJNI = null;
-                            while (flutterJNI == null) {
-                                flutterJNI = getFlutterJNIReference();
-                                flutterJNI.markTextureFrameAvailable(id);
-                            }
-                        } catch (Throwable e) {
-                            e.printStackTrace();
-                        }
-                    }
-                }, new Handler());
-            } else {
-                surfaceTextureEntry.surfaceTexture().setOnFrameAvailableListener((texture) -> {
-                    synchronized (lock) {
-                        try {
-                            if (!waitUntilFirstFrameRenderedNotify) {
-                                waitUntilFirstFrameRenderedNotify = true;
-                                final HashMap<String, Object> data = new HashMap<>();
-                                data.put("handle", handle);
-                                channelReference.invokeMethod("VideoOutput.WaitUntilFirstFrameRenderedNotify", data);
-                                Log.i("media_kit", String.format(Locale.ENGLISH, "VideoOutput.WaitUntilFirstFrameRenderedNotify = %d", handle));
-                            }
+    private final TextureRegistry.SurfaceProducer surfaceProducer;
 
-                            FlutterJNI flutterJNI = null;
-                            while (flutterJNI == null) {
-                                flutterJNI = getFlutterJNIReference();
-                                flutterJNI.markTextureFrameAvailable(id);
-                            }
-                        } catch (Throwable e) {
-                            e.printStackTrace();
-                        }
-                    }
-                });
-            }
-        } else {
-            if (!waitUntilFirstFrameRenderedNotify) {
-                waitUntilFirstFrameRenderedNotify = true;
-                final HashMap<String, Object> data = new HashMap<>();
-                data.put("id", id);
-                data.put("wid", wid);
-                data.put("handle", handle);
-                channelReference.invokeMethod("VideoOutput.WaitUntilFirstFrameRenderedNotify", data);
-            }
-        }
+    private final Object lock = new Object();
 
-        try {
-            id = surfaceTextureEntry.id();
-            Log.i("media_kit", String.format(Locale.ENGLISH, "com.alexmercerind.media_kit_video.VideoOutput: id = %d", id));
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
+    VideoOutput(TextureRegistry textureRegistryReference, TextureUpdateCallback textureUpdateCallback) {
+        this.textureUpdateCallback = textureUpdateCallback;
+
+        surfaceProducer = textureRegistryReference.createSurfaceProducer();
+        surfaceProducer.setCallback(this);
+
+        // By default, android.graphics.SurfaceTexture has a size of 1x1.
+        setSurfaceSize(1, 1, true);
     }
 
     public void dispose() {
-        try {
-            surfaceTextureEntry.release();
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
-        try {
-            surface.release();
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
-        try {
-            final Handler handler = new Handler(Looper.getMainLooper());
-            handler.postDelayed(() -> {
-                try {
-                    // Invoke DeleteGlobalRef after a voluntary delay to eliminate possibility of libmpv referencing it sometime in the near future.
-                    deleteGlobalObjectRef.invoke(null, wid);
-                    Log.i("media_kit", String.format(Locale.ENGLISH, "com.alexmercerind.mediakitandroidhelper.MediaKitAndroidHelper.deleteGlobalObjectRef: %d", wid));
-                } catch (Throwable e) {
-                    e.printStackTrace();
-                }
-            }, 5000);
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
-    }
-
-    public long createSurface() {
         synchronized (lock) {
-            // Delete previous android.view.Surface & object reference.
             try {
-                if (surface != null) {
-                    clearSurface();
-                    surface.release();
-                    surface = null;
-                }
-                if (wid != 0) {
-                    deleteGlobalObjectRef.invoke(null, wid);
-                    wid = 0;
-                }
+                surfaceProducer.getSurface().release();
             } catch (Throwable e) {
-                e.printStackTrace();
+                Log.e(TAG, "dispose", e);
             }
-            // Create new android.view.Surface & object reference.
             try {
-                surface = new Surface(surfaceTextureEntry.surfaceTexture());
-                wid = (long) newGlobalObjectRef.invoke(null, surface);
+                surfaceProducer.release();
             } catch (Throwable e) {
-                e.printStackTrace();
+                Log.e(TAG, "dispose", e);
             }
-            return wid;
+            onSurfaceDestroyed();
         }
     }
 
-    public void setSurfaceTextureSize(int width, int height) {
-        try {
-            surfaceTextureEntry.surfaceTexture().setDefaultBufferSize(width, height);
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
+    public void setSurfaceSize(int width, int height) {
+        setSurfaceSize(width, height, false);
     }
 
-    private void clearSurface() {
-        try {
-            final Canvas canvas = surface.lockCanvas(null);
-            canvas.drawColor(Color.TRANSPARENT, PorterDuff.Mode.CLEAR);
-            surface.unlockCanvasAndPost(canvas);
-        } catch (Throwable e) {
-            e.printStackTrace();
-        }
-    }
-
-    private FlutterJNI getFlutterJNIReference() {
-        try {
-            FlutterView view = null;
-            // io.flutter.embedding.android.FlutterActivity
-            if (view == null) {
-                view = MediaKitVideoPlugin.activity.findViewById(FlutterActivity.FLUTTER_VIEW_ID);
-            }
-            // io.flutter.embedding.android.FlutterFragmentActivity
-            if (view == null) {
-                final FrameLayout layout = (FrameLayout) MediaKitVideoPlugin.activity.findViewById(FlutterFragmentActivity.FRAGMENT_CONTAINER_ID);
-                for (int i = 0; i < layout.getChildCount(); i++) {
-                    final View child = layout.getChildAt(i);
-                    if (child instanceof FlutterView) {
-                        view = (FlutterView) child;
-                        break;
-                    }
+    private void setSurfaceSize(int width, int height, boolean force) {
+        synchronized (lock) {
+            try {
+                if (!force && surfaceProducer.getWidth() == width && surfaceProducer.getHeight() == height) {
+                    return;
                 }
+                surfaceProducer.setSize(width, height);
+                onSurfaceCreated();
+            } catch (Throwable e) {
+                Log.e(TAG, "setSurfaceSize", e);
             }
-            final FlutterEngine engine = view.getAttachedFlutterEngine();
-            final Field field = engine.getClass().getDeclaredField("flutterJNI");
-            field.setAccessible(true);
-            return (FlutterJNI) field.get(engine);
+        }
+    }
+
+    @Override
+    public void onSurfaceCreated() {
+        synchronized (lock) {
+            Log.i(TAG, "onSurfaceCreated");
+            id = surfaceProducer.id();
+            wid = newGlobalObjectRef(surfaceProducer.getSurface());
+            textureUpdateCallback.onTextureUpdate(id, wid, surfaceProducer.getWidth(), surfaceProducer.getHeight());
+        }
+    }
+
+    @Override
+    public void onSurfaceDestroyed() {
+        synchronized (lock) {
+            Log.i(TAG, "onSurfaceDestroyed");
+            textureUpdateCallback.onTextureUpdate(id, 0, surfaceProducer.getWidth(), surfaceProducer.getHeight());
+            if (wid != 0) {
+                final long widReference = wid;
+                handler.postDelayed(() -> deleteGlobalObjectRef(widReference), 5000);
+            }
+        }
+    }
+
+    private static long newGlobalObjectRef(Object object) {
+        Log.i(TAG, String.format(Locale.ENGLISH, "newGlobalRef: object = %s", object));
+        try {
+            return (long) Objects.requireNonNull(newGlobalObjectRef.invoke(null, object));
         } catch (Throwable e) {
-            e.printStackTrace();
-            return null;
+            Log.e(TAG, "newGlobalRef", e);
+            return 0;
+        }
+    }
+
+    private static void deleteGlobalObjectRef(long ref) {
+        if (deletedGlobalObjectRefs.contains(ref)) {
+            Log.i(TAG, String.format(Locale.ENGLISH, "deleteGlobalObjectRef: ref = %d ALREADY DELETED", ref));
+            return;
+        }
+        if (deletedGlobalObjectRefs.size() > 100) {
+            deletedGlobalObjectRefs.clear();
+        }
+        deletedGlobalObjectRefs.add(ref);
+        Log.i(TAG, String.format(Locale.ENGLISH, "deleteGlobalObjectRef: ref = %d", ref));
+        try {
+            deleteGlobalObjectRef.invoke(null, ref);
+        } catch (Throwable e) {
+            Log.e(TAG, "deleteGlobalObjectRef", e);
         }
     }
 }
