@@ -4,21 +4,15 @@
 /// All rights reserved.
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 import 'dart:io';
-import 'dart:ffi';
 import 'dart:async';
 import 'dart:collection';
 import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
+import 'package:synchronized/synchronized.dart';
 
 import 'package:media_kit/media_kit.dart';
-// ignore_for_file: unused_import, implementation_imports
-import 'package:media_kit/ffi/ffi.dart';
-import 'package:media_kit/src/player/native/core/native_library.dart';
-
-import 'package:media_kit/generated/libmpv/bindings.dart';
 
 import 'package:media_kit_video/src/utils/query_decoders.dart';
-import 'package:media_kit_video/src/video_controller/video_controller.dart';
 import 'package:media_kit_video/src/video_controller/platform_video_controller.dart';
 
 /// {@template native_video_controller}
@@ -47,18 +41,86 @@ class NativeVideoController extends PlatformVideoController {
   /// Fixed height of the video output.
   int? height;
 
+  /// Width of the video (from [VideoParams]).
+  int? videoParamsWidth;
+
+  /// Height of the video (from [VideoParams]).
+  int? videoParamsHeight;
+
+  /// [Lock] used to synchronize [onLoadHooks], [onUnloadHooks] & [subscription].
+  final lock = Lock();
+
+  NativePlayer get platform => player.platform as NativePlayer;
+
+  Future<void> setProperty(String key, String value) async {
+    await platform.setProperty(key, value, waitForInitialization: false);
+  }
+
+  Future<void> setProperties(Map<String, String> properties) async {
+    // ORDER IS IMPORTANT.
+    for (final entry in properties.entries) {
+      await setProperty(entry.key, entry.value);
+    }
+  }
+
+  /// [StreamSubscription] for listening to video [Rect].
+  StreamSubscription<VideoParams>? videoParamsSubscription;
+
   /// {@macro native_video_controller}
   NativeVideoController._(
     super.player,
     super.configuration,
   )   : width = configuration.width,
-        height = configuration.height;
+        height = configuration.height {
+    videoParamsSubscription = player.stream.videoParams.listen(
+      (event) => lock.synchronized(() async {
+        if ([0, null].contains(event.dw) || [0, null].contains(event.dh)) {
+          return;
+        }
+
+        final int handle = await player.handle;
+
+        final int width;
+        final int height;
+        if (event.rotate == 0 || event.rotate == 180) {
+          width = event.dw ?? 0;
+          height = event.dh ?? 0;
+        } else {
+          // width & height are swapped for 90 or 270 degrees rotation.
+          width = event.dh ?? 0;
+          height = event.dw ?? 0;
+        }
+
+        if (videoParamsWidth == width && videoParamsHeight == height) {
+          return;
+        }
+
+        videoParamsWidth = width;
+        videoParamsHeight = height;
+
+        await _channel.invokeMethod(
+          'VideoOutputManager.SetSize',
+          {
+            'handle': handle.toString(),
+            'width': width.toString(),
+            'height': height.toString(),
+          },
+        );
+      }),
+    );
+  }
 
   /// {@macro native_video_controller}
   static Future<PlatformVideoController> create(
     Player player,
     VideoControllerConfiguration configuration,
   ) async {
+    // Update [configuration] to have default values.
+    configuration = configuration.copyWith(
+      vo: configuration.vo ?? 'libmpv',
+      hwdec: configuration.hwdec ?? 'auto',
+    );
+
     // Retrieve the native handle of the [Player].
     final handle = await player.handle;
     // Return the existing [VideoController] if it's already created.
@@ -91,34 +153,21 @@ class NativeVideoController extends PlatformVideoController {
     // Store the [NativeVideoController] in the [_controllers].
     _controllers[handle] = controller;
 
-    // ----------------------------------------------
-    NativeLibrary.ensureInitialized();
-    final mpv = MPV(DynamicLibrary.open(NativeLibrary.path));
-    final values = {
-      'vo': configuration.vo ?? 'libmpv',
-      'hwdec': configuration.hwdec ?? 'auto',
-      'vid': 'auto',
-    };
-    for (final entry in values.entries) {
-      final property = entry.key.toNativeUtf8();
-      final value = entry.value.toNativeUtf8();
-      mpv.mpv_set_property_string(
-        Pointer.fromAddress(handle),
-        property.cast(),
-        value.cast(),
-      );
-      calloc.free(property);
-      calloc.free(value);
-    }
-    // ----------------------------------------------
+    await controller.setProperties(
+      {
+        'vo': configuration.vo!,
+        'hwdec': configuration.hwdec!,
+        'vid': 'auto',
+      },
+    );
 
-    // Wait until first texture ID is received i.e. render context & EGL/D3D surface is created.
+    // Wait until first texture ID is received.
     // We are not waiting on the native-side itself because it will block the UI thread.
-    // Background platform channels are not a thing yet.
     final completer = Completer<void>();
     void listener() {
-      if (controller.id.value != null) {
-        debugPrint('NativeVideoController: Texture ID: ${controller.id.value}');
+      final value = controller.id.value;
+      if (value != null) {
+        debugPrint('NativeVideoController: Texture ID: $value');
         completer.complete();
       }
     }
@@ -179,8 +228,8 @@ class NativeVideoController extends PlatformVideoController {
         'VideoOutputManager.SetSize',
         {
           'handle': handle.toString(),
-          'width': 'null',
-          'height': 'null',
+          'width': videoParamsWidth?.toString() ?? 'null',
+          'height': videoParamsHeight?.toString() ?? 'null',
         },
       );
     }
@@ -188,6 +237,8 @@ class NativeVideoController extends PlatformVideoController {
 
   /// Disposes the instance. Releases allocated resources back to the system.
   Future<void> _dispose() async {
+    super.dispose();
+    await videoParamsSubscription?.cancel();
     final handle = await player.handle;
     _controllers.remove(handle);
     await _channel.invokeMethod(
