@@ -4,8 +4,7 @@
 /// All rights reserved.
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 import 'dart:async';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
 import 'package:media_kit/media_kit.dart';
 
 import 'package:media_kit_video/src/video_controller/platform_video_controller.dart';
@@ -78,6 +77,10 @@ class VideoController {
     player.platform?.isVideoControllerAttached = true;
 
     () async {
+      final completer = Completer();
+      WidgetsBinding.instance.addPostFrameCallback((_) => completer.complete());
+      await completer.future;
+
       try {
         if (NativeVideoController.supported) {
           final result = await NativeVideoController.create(
