@@ -8,6 +8,11 @@
 
 #include "utils.h"
 
+typedef LONG NTSTATUS, *PNTSTATUS;
+#define STATUS_SUCCESS (0x00000000)
+
+typedef NTSTATUS(WINAPI* RtlGetVersionPtr)(PRTL_OSVERSIONINFOW);
+
 void Utils::EnterNativeFullscreen(HWND window) {
   if (fullscreen_) {
     return;
@@ -38,7 +43,8 @@ void Utils::EnterNativeFullscreen(HWND window) {
                      &monitor);
     ::SetWindowLongPtr(window, GWL_STYLE, style & ~WS_OVERLAPPEDWINDOW);
     ::SetWindowPos(window, HWND_TOP, monitor.rcMonitor.left,
-                   monitor.rcMonitor.top, monitor.rcMonitor.right - monitor.rcMonitor.left,
+                   monitor.rcMonitor.top,
+                   monitor.rcMonitor.right - monitor.rcMonitor.left,
                    monitor.rcMonitor.bottom - monitor.rcMonitor.top,
                    SWP_NOOWNERZORDER | SWP_FRAMECHANGED);
   }
@@ -74,6 +80,26 @@ void Utils::ExitNativeFullscreen(HWND window) {
           SWP_NOACTIVATE | SWP_NOZORDER);
     }
   }
+}
+
+RTL_OSVERSIONINFOW Utils::GetWindowsVersion() {
+  HMODULE handle = ::LoadLibraryW(L"ntdll.dll");
+  RTL_OSVERSIONINFOW rtl_os_version_info = {0};
+  rtl_os_version_info.dwBuildNumber = 0;
+  rtl_os_version_info.dwOSVersionInfoSize = sizeof(rtl_os_version_info);
+  if (handle) {
+    RtlGetVersionPtr rtl_get_version_ptr = reinterpret_cast<RtlGetVersionPtr>(
+        ::GetProcAddress(handle, "RtlGetVersion"));
+    if (rtl_get_version_ptr != nullptr) {
+      rtl_get_version_ptr(&rtl_os_version_info);
+    }
+    ::FreeLibrary(handle);
+  }
+  return rtl_os_version_info;
+}
+
+bool Utils::IsWindows10RTMOrGreater() {
+  return GetWindowsVersion().dwBuildNumber >= 10240;
 }
 
 bool Utils::fullscreen_ = false;
